+++ src/main.rs
@@ -1,3 +1,4 @@
+use clap::{arg, command};
 use parser::constraints_from_json;
 
 pub mod parser;
@@ -9,7 +10,10 @@ use solver::Solver;
 
 fn main() {
     env_logger::init();
-    let program = constraints_from_json("tests/retypd-constrains-simple.json").unwrap();
+    let matches = command!()
+        .arg(arg!([json_in] "Path to the constraints json file").default_value("tests/retypd-constrains-simple.json"))
+        .get_matches();
+    let program = constraints_from_json(matches.get_one::<String>("json_in").unwrap()).unwrap();
     let solver = Solver::new(&program);
     solver.infer_shapes();
 }