+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ### TODO
 
-1. 
+1. add commandline. OK
+   1. path to print the debug graph. OK
+2. implement the initial graph build.